Specify request options when querying HDFS ACLs
The request_options dictionary contains a set of extra options used to instantiate the Tornado httpclient.HTTPRequest class. The request_options are defined once in the constructor of the webhdfs client and the same value is used anytime the HTTPRequest class is instantiated except for one call. 
This commit adds the request_options to the missed request so that globally defined request options are passed to all HttpRequests created in this file.